finish part 5 end to end testing

+++ Part5/Blogs/blog-backend/controllers/blogs.js
@@ -1,35 +1,91 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  // const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
-
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
+  // const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
 
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0
-  })
+  // This is used when the tokenExtractor middleware is used
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-  const savedBlog = await blog.save()
-  response.status(201).json(savedBlog)
-})
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
 
-blogsRouter.delete('/:id', (request, response, next) => {
-  Blog.findByIdAndDelete(request.params.id)
-    .then(() => {
-      response.status(204).end()
+  // This is used when the userExtractor middleware is used
+  try {  
+    const user = request.user
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user._id
     })
-    .catch(error => next(error))
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    const populatedBlog = await savedBlog.populate('user', { username: 1, name: 1 })
+    response.status(201).json(populatedBlog)
+  } catch (error) {
+    response.status(401).json({ error: `Unauthorized. Blog creation failed. Reason: ${error}` })
+  }
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog) {
+    response.json(blog)
+  } else {
+    response.status(404).end()
+  }
+})
+
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
+
+  // This is used when the "userExtractor" middleware is used....
+  const user = request.user
+
+  // ...this is when "userExtractor" NOT used 
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() == user._id.toString()) {
+    Blog.findByIdAndDelete(request.params.id)
+      .then(() => {
+        response.status(204).end()
+      })
+      .catch(error => next(error))
+    return response.status(200).json({ message: 'Blog deleted successfully' })
+  }
+  return response.status(403).json({ error: 'Only the creator can delete this blog' })
+
+})
+
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
   const body = request.body
 
   const blog = {
